+++ src/controller_rpc.c
@@ -3807,7 +3807,7 @@ controller_edit_config(clixon_handle h,
     cxobj     *xserv;
     int        ret;
 
-    clixon_debug(CLIXON_DBG_CTRL, "wrapper");
+    clixon_debug(CLIXON_DBG_CTRL, "Find and remove creator attributes and create services/../created structures");
     if ((yspec = clicon_dbspec_yang(h)) == NULL){
         clixon_err(OE_YANG, ENOENT, "No yang spec9");
         goto done;
@@ -3849,7 +3849,7 @@ controller_edit_config(clixon_handle h,
     }
     if (xml_child_nr_type(xserv, CX_ELMNT) == 0)
         goto ok;
-    clixon_debug_xml(CLIXON_DBG_CTRL, xserv, "Objects created in %s-db", target);
+    clixon_debug_xml(CLIXON_DBG_CTRL2, xserv, "Objects created in %s-db", target);
     if ((ret = xmldb_put(h, target, OP_NONE, xconfig, NULL, cbret)) < 0){
         if (netconf_operation_failed(cbret, "protocol", "%s", clixon_err_reason())< 0)
             goto done;